feat: update career details page to new layout
