+++ src/components/career-form/career-form.tsx
@@ -204,7 +204,7 @@ export const CareerForm: React.FC<CareerFormProps> = (props) => {
   };
 
   const onError = (event) => {
-    console.log('onError', event);
+    // console.log('onError', event);
   };
 
   // recover from some previous error
